requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.7.2
async-timeout==3.0.1
attrs==21.2.0
black==21.7b0
cachetools==4.2.2
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.4
ciso8601==2.2.0
click==8.0.1
dateparser==1.0.0
decorator==5.0.9
discord-webhook==0.14.0
ftx==1.0.2
idna==2.10
iniconfig==1.1.1
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
packaging==21.0
pathlib==1.0.1
pathspec==0.9.0
platformdirs==2.2.0
pluggy==1.0.0
py==1.10.0
pydantic==1.8.2
pylint==2.10.2
pyparsing==2.4.7
pytest==6.2.5
python-binance==1.0.12
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.23.0
satoshi==0.1.3
schedule==0.6.0
six==1.16.0
toml==0.10.2
tomli==1.2.1
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==3.0
ujson==4.1.0
urllib3==1.25.11
websockets==9.1
wrapt==1.12.1
yarl==1.6.3
